Moved a lot of things around which broke a lot of things. Moving weapon and armor qualities to their own components.
