+++ bodies/blood.py
@@ -1,0 +1,7 @@
+class Blood(object):
+    """
+    Most characters will have ordinary blood but some could have acidic blood or with other properties.
+    """
+    def __init__(self, uid, name):
+        self.uid = uid
+        self.name = name
