+++ characters/character.py
@@ -1,11 +1,11 @@
 import math
 
 from characters.enums import Sex
-from components.location import Location
-from components.equipment import Equipment
-from components.stats import Stats
 from combat.attack import base_attacks
 from combat.defense import base_defenses
+from components.equipment import Equipment
+from components.location import Location
+from components.stats import Stats
 
 
 class Character(object):
