+++ components/component.py
@@ -1,0 +1,5 @@
+class Component(object):
+    """
+    Base class for components, no use right now.
+    """
+    pass
