+++ components/display.py
@@ -1,4 +1,7 @@
-class Display(object):
+from .component import Component
+
+
+class Display(Component):
     def __init__(self, foreground_color, background_color, ascii_character):
         self.foreground_color = foreground_color
         self.background_color = background_color
