--- components/game_object.py
@@ -1,5 +1,0 @@
-# TODO This is to follow a pattern used in XNA, MONOGAME OR UNITY
-# TODO It is a BASE class that can be shared by ANYTHING that has a presence ingame.
-# TODO Walls, items, characters, fireballs
-# TODO In python it would only serve as a "helper" than a necessity.
-# TODO I know it uses duck typing for the most part but I think it will help long term.
