+++ data/json_template_loader.py
@@ -2,8 +2,12 @@ import logging
 import jsonpickle
 import os
 from components.material import materials
-from components.body import get_humanoid_body_sample, get_body_parts_sample
-from characters.samples import get_sample_monsters, get_sample_classes, get_race_samples, get_sample_items
+from data.python_templates.body_parts import body_parts
+from data.python_templates.body import bodies
+from data.python_templates.items import items
+from data.python_templates.characters import characters
+from data.python_templates.classes import character_classes
+from data.python_templates.races import races
 
 logger_ = logging.getLogger()
 
@@ -63,10 +67,10 @@ class JsonTemplateManager(object):
 if __name__ == '__main__':
     template_manager = JsonTemplateManager()
     template_manager.material_templates = {material.uid: material for material in materials}
-    template_manager.bodies_templates = {body.uid: body for body in get_humanoid_body_sample()}
-    template_manager.bodyparts_templates = {body_part.uid: body_part for body_part in get_body_parts_sample()}
-    template_manager.race_templates = {race.uid: race for race in get_race_samples()}
-    template_manager.class_templates = {c_class.uid: c_class for c_class in get_sample_classes()}
-    template_manager.monster_templates = {monster.uid: monster for monster in get_sample_monsters()}
-    template_manager.item_templates = {item.uid: item for item in get_sample_items()}
+    template_manager.bodies_templates = {body.uid: body for body in bodies}
+    template_manager.bodyparts_templates = {body_part.uid: body_part for body_part in body_parts}
+    template_manager.race_templates = {race.uid: race for race in races}
+    template_manager.class_templates = {c_class.uid: c_class for c_class in character_classes}
+    template_manager.monster_templates = {monster.uid: monster for monster in characters}
+    template_manager.item_templates = {item.uid: item for item in items}
     template_manager.save_templates()
