+++ data/templates/bodies.json
@@ -1,1 +1,1 @@
-{"humanoid": {"blood_type": "humanoid_blood", "structural_material": "bone", "py/object": "components.body.Body", "outer_material_uid": null, "uid": "humanoid", "bodypart_tree": {"py/object": "components.body.BodypartTree", "nodes": [{"name": "torso", "body_part_uid": "humanoid_torso", "py/object": "components.body.BodypartTreeNode", "children_nodes": [{"name": "heart", "body_part_uid": "humanoid_heart", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 2, "instance": null}, {"name": "lungs", "body_part_uid": "humanoid_lungs", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 2, "instance": null}, {"name": "head", "body_part_uid": "humanoid_head", "py/object": "components.body.BodypartTreeNode", "children_nodes": [{"name": "left eye", "body_part_uid": "humanoid_eye", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}, {"name": "right eye", "body_part_uid": "humanoid_eye", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}, {"name": "left ear", "body_part_uid": "humanoid_ear", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}, {"name": "right ear", "body_part_uid": "humanoid_ear", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}, {"name": "mouth", "body_part_uid": "humanoid_mouth", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}, {"name": "brain", "body_part_uid": "humanoid_brain", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 2, "instance": null}], "connection_type": 1, "instance": null}, {"name": "left arm", "body_part_uid": "humanoid_arm", "py/object": "components.body.BodypartTreeNode", "children_nodes": [{"name": "left hand", "body_part_uid": "humanoid_hand", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}], "connection_type": 1, "instance": null}, {"name": "right arm", "body_part_uid": "humanoid_arm", "py/object": "components.body.BodypartTreeNode", "children_nodes": [{"name": "right hand", "body_part_uid": "humanoid_hand", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}], "connection_type": 1, "instance": null}, {"name": "left leg", "body_part_uid": "humanoid_leg", "py/object": "components.body.BodypartTreeNode", "children_nodes": [{"name": "left foot", "body_part_uid": "humanoid_foot", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}], "connection_type": 1, "instance": null}, {"name": "right leg", "body_part_uid": "humanoid_leg", "py/object": "components.body.BodypartTreeNode", "children_nodes": [{"name": "right foot", "body_part_uid": "humanoid_foot", "py/object": "components.body.BodypartTreeNode", "children_nodes": [], "connection_type": 1, "instance": null}], "connection_type": 1, "instance": null}], "connection_type": 0, "instance": null}, {"py/id": 5}, {"py/id": 7}, {"py/id": 9}, {"py/id": 11}, {"py/id": 13}, {"py/id": 15}, {"py/id": 17}, {"py/id": 19}, {"py/id": 21}, {"py/id": 23}, {"py/id": 25}, {"py/id": 27}, {"py/id": 29}, {"py/id": 31}, {"py/id": 33}, {"py/id": 35}, {"py/id": 37}]}, "inner_material_uid": null, "structural_material_uid": null, "blood_uid": null, "name": "Humanoid", "inner_material": "flesh", "height": 5, "weight": 150, "outer_material": "skin"}}
+{"humanoid": {"structural_material_uid": null, "height": 5, "weight": 150, "uid": "humanoid", "inner_material_uid": null, "inner_material": "flesh", "outer_material": "skin", "outer_material_uid": null, "blood_uid": null, "py/object": "components.body.Body", "structural_material": "bone", "blood_type": "humanoid_blood", "name": "Humanoid", "bodypart_tree": {"py/object": "bodies.body_part_tree.BodypartTree", "nodes": [{"connection_type": 0, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "torso", "body_part_uid": "humanoid_torso", "children_nodes": [{"connection_type": 2, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "heart", "body_part_uid": "humanoid_heart", "children_nodes": []}, {"connection_type": 2, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "lungs", "body_part_uid": "humanoid_lungs", "children_nodes": []}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "head", "body_part_uid": "humanoid_head", "children_nodes": [{"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "left eye", "body_part_uid": "humanoid_eye", "children_nodes": []}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "right eye", "body_part_uid": "humanoid_eye", "children_nodes": []}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "left ear", "body_part_uid": "humanoid_ear", "children_nodes": []}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "right ear", "body_part_uid": "humanoid_ear", "children_nodes": []}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "mouth", "body_part_uid": "humanoid_mouth", "children_nodes": []}, {"connection_type": 2, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "brain", "body_part_uid": "humanoid_brain", "children_nodes": []}]}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "left arm", "body_part_uid": "humanoid_arm", "children_nodes": [{"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "left hand", "body_part_uid": "humanoid_hand", "children_nodes": []}]}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "right arm", "body_part_uid": "humanoid_arm", "children_nodes": [{"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "right hand", "body_part_uid": "humanoid_hand", "children_nodes": []}]}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "left leg", "body_part_uid": "humanoid_leg", "children_nodes": [{"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "left foot", "body_part_uid": "humanoid_foot", "children_nodes": []}]}, {"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "right leg", "body_part_uid": "humanoid_leg", "children_nodes": [{"connection_type": 1, "instance": null, "py/object": "bodies.body_part_tree.BodypartTreeNode", "name": "right foot", "body_part_uid": "humanoid_foot", "children_nodes": []}]}]}, {"py/id": 5}, {"py/id": 7}, {"py/id": 9}, {"py/id": 11}, {"py/id": 13}, {"py/id": 15}, {"py/id": 17}, {"py/id": 19}, {"py/id": 21}, {"py/id": 23}, {"py/id": 25}, {"py/id": 27}, {"py/id": 29}, {"py/id": 31}, {"py/id": 33}, {"py/id": 35}, {"py/id": 37}]}}}
