+++ data/templates/bodyparts.json
@@ -1,1 +1,1 @@
-{"humanoid_mouth": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.EAT: 7>": 1}, "threat_level": {"py/id": 5}, "relative_size": 5, "uid": "humanoid_mouth"}, "humanoid_torso": {"py/object": "components.body.BodyPart", "physical_abilities": {}, "threat_level": {"py/id": 1}, "relative_size": 50, "uid": "humanoid_torso"}, "humanoid_foot": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.MOVE: 9>": 1, "<PhysicalAbilities.STAND: 1>": 1}, "threat_level": {"py/id": 5}, "relative_size": 10, "uid": "humanoid_foot"}, "humanoid_brain": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.THINK: 8>": 1}, "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [3]}, null, null, null]}, "relative_size": 15, "uid": "humanoid_brain"}, "humanoid_arm": {"py/object": "components.body.BodyPart", "physical_abilities": {}, "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [1]}, null, null, null]}, "relative_size": 25, "uid": "humanoid_arm"}, "humanoid_ear": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.HEAR: 3>": 1}, "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [0]}, null, null, null]}, "relative_size": 5, "uid": "humanoid_ear"}, "humanoid_heart": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.LIVE: 10>": 1}, "threat_level": {"py/id": 3}, "relative_size": 25, "uid": "humanoid_heart"}, "humanoid_head": {"py/object": "components.body.BodyPart", "physical_abilities": {}, "threat_level": {"py/id": 1}, "relative_size": 25, "uid": "humanoid_head"}, "humanoid_lungs": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.BREATHE: 11>": 1}, "threat_level": {"py/id": 3}, "relative_size": 25, "uid": "humanoid_lungs"}, "humanoid_hand": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.GRASP: 0>": 1, "<PhysicalAbilities.PUNCH: 5>": 1}, "threat_level": {"py/id": 5}, "relative_size": 10, "uid": "humanoid_hand"}, "humanoid_leg": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.MOVE: 9>": 1, "<PhysicalAbilities.STAND: 1>": 1}, "threat_level": {"py/id": 1}, "relative_size": 25, "uid": "humanoid_leg"}, "humanoid_eye": {"py/object": "components.body.BodyPart", "physical_abilities": {"<PhysicalAbilities.SEE: 2>": 1}, "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [2]}, null, null, null]}, "relative_size": 5, "uid": "humanoid_eye"}}
+{"humanoid_foot": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.STAND: 1>": 1, "<PhysicalAbilities.MOVE: 9>": 1}, "uid": "humanoid_foot", "threat_level": {"py/id": 5}, "relative_size": 10}, "humanoid_brain": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.THINK: 8>": 1}, "uid": "humanoid_brain", "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [3]}, null, null, null]}, "relative_size": 15}, "humanoid_lungs": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.BREATHE: 11>": 1}, "uid": "humanoid_lungs", "threat_level": {"py/id": 3}, "relative_size": 25}, "humanoid_leg": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.STAND: 1>": 1, "<PhysicalAbilities.MOVE: 9>": 1}, "uid": "humanoid_leg", "threat_level": {"py/id": 1}, "relative_size": 25}, "humanoid_heart": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.LIVE: 10>": 1}, "uid": "humanoid_heart", "threat_level": {"py/id": 3}, "relative_size": 25}, "humanoid_ear": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.HEAR: 3>": 1}, "uid": "humanoid_ear", "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [0]}, null, null, null]}, "relative_size": 5}, "humanoid_hand": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.PUNCH: 5>": 1, "<PhysicalAbilities.GRASP: 0>": 1}, "uid": "humanoid_hand", "threat_level": {"py/id": 5}, "relative_size": 10}, "humanoid_eye": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.SEE: 2>": 1}, "uid": "humanoid_eye", "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [2]}, null, null, null]}, "relative_size": 5}, "humanoid_arm": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {}, "uid": "humanoid_arm", "threat_level": {"py/object": "combat.enums.ThreatLevel", "py/reduce": [{"py/type": "combat.enums.ThreatLevel"}, {"py/tuple": [1]}, null, null, null]}, "relative_size": 25}, "humanoid_head": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {}, "uid": "humanoid_head", "threat_level": {"py/id": 1}, "relative_size": 25}, "humanoid_mouth": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {"<PhysicalAbilities.EAT: 7>": 1}, "uid": "humanoid_mouth", "threat_level": {"py/id": 5}, "relative_size": 5}, "humanoid_torso": {"py/object": "bodies.body_part.BodyPart", "physical_abilities": {}, "uid": "humanoid_torso", "threat_level": {"py/id": 1}, "relative_size": 50}}
