+++ data/templates/classes.json
@@ -1,1 +1,1 @@
-{"warrior": {"py/object": "characters.classes.CharacterClass", "experience_penalty": 0, "uid": "warrior", "name": "Warrior", "level_tree": {"py/object": "components.level_tree.LevelTree", "abilities_modifiers": {"2": {"<PowerAbilities.PowerAttack: 2>": 1}, "4": {"<PowerAbilities.Parry: 3>": 1}}, "stats_modifiers": {"0": {"<Stats.Health: 'health'>": {"py/object": "components.stats.StatModifier", "uid": {"py/id": 3}, "value": 10, "level_progression": 1}}}}}, "thief": {"py/object": "characters.classes.CharacterClass", "experience_penalty": 0, "uid": "thief", "name": "Thief", "level_tree": {"py/object": "components.level_tree.LevelTree", "abilities_modifiers": {"2": {"<PowerAbilities.Sneak: 4>": 1}, "4": {"<PowerAbilities.Backstab: 5>": 1}}, "stats_modifiers": {"0": {"<Stats.Health: 'health'>": {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["health"]}, null, null, null]}, "value": 6, "level_progression": 1}}}}}}
+{"warrior": {"py/object": "characters.classes.CharacterClass", "experience_penalty": 0, "uid": "warrior", "name": "Warrior", "level_tree": {"py/object": "components.level_tree.LevelTree", "stats_modifiers": {"0": {"<Stats.Health: 'health'>": {"py/object": "components.stats.StatModifier", "uid": {"py/id": 3}, "level_progression": 1, "value": 10}}}, "abilities_modifiers": {"2": {"<PowerAbilities.PowerAttack: 2>": 1}, "4": {"<PowerAbilities.Parry: 3>": 1}}}}, "thief": {"py/object": "characters.classes.CharacterClass", "experience_penalty": 0, "uid": "thief", "name": "Thief", "level_tree": {"py/object": "components.level_tree.LevelTree", "stats_modifiers": {"0": {"<Stats.Health: 'health'>": {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["health"]}, null, null, null]}, "level_progression": 1, "value": 6}}}, "abilities_modifiers": {"2": {"<PowerAbilities.Sneak: 4>": 1}, "4": {"<PowerAbilities.Backstab: 5>": 1}}}}}
