+++ data/templates/materials.json
@@ -1,1 +1,1 @@
-{"bone": {"hardness": 1, "uid": "bone", "name": "Bone", "value": 0.1, "py/object": "components.material.Material", "potency": 1, "weight": 1, "sharpness": 1}, "iron": {"hardness": 2, "uid": "iron", "name": "Iron", "value": 1, "py/object": "components.material.Material", "potency": 1, "weight": 2, "sharpness": 2}, "skin": {"hardness": 0.1, "uid": "skin", "name": "Skin", "value": 0.1, "py/object": "components.material.Material", "potency": 1, "weight": 0.1, "sharpness": 0}, "flesh": {"hardness": 0.2, "uid": "flesh", "name": "Flesh", "value": 0.1, "py/object": "components.material.Material", "potency": 1, "weight": 0.2, "sharpness": 0}, "stone": {"hardness": 2, "uid": "stone", "name": "Stone", "value": 0.5, "py/object": "components.material.Material", "potency": 0.5, "weight": 2, "sharpness": 1}, "wood": {"hardness": 1, "uid": "wood", "name": "Wood", "value": 1, "py/object": "components.material.Material", "potency": 1, "weight": 0.5, "sharpness": 0.5}}
+{"skin": {"value": 0.1, "weight": 0.1, "potency": 1, "hardness": 0.1, "py/object": "components.material.Material", "uid": "skin", "name": "Skin", "sharpness": 0}, "iron": {"value": 1, "weight": 2, "potency": 1, "hardness": 2, "py/object": "components.material.Material", "uid": "iron", "name": "Iron", "sharpness": 2}, "bone": {"value": 0.1, "weight": 1, "potency": 1, "hardness": 1, "py/object": "components.material.Material", "uid": "bone", "name": "Bone", "sharpness": 1}, "flesh": {"value": 0.1, "weight": 0.2, "potency": 1, "hardness": 0.2, "py/object": "components.material.Material", "uid": "flesh", "name": "Flesh", "sharpness": 0}, "stone": {"value": 0.5, "weight": 2, "potency": 0.5, "hardness": 2, "py/object": "components.material.Material", "uid": "stone", "name": "Stone", "sharpness": 1}, "wood": {"value": 1, "weight": 0.5, "potency": 1, "hardness": 1, "py/object": "components.material.Material", "uid": "wood", "name": "Wood", "sharpness": 0.5}}
