+++ data/templates/races.json
@@ -1,1 +1,1 @@
-{"human": {"experience_penalty": 0, "uid": "human", "name": "Human", "level_tree": {"py/object": "components.level_tree.LevelTree", "abilities_modifiers": {}, "stats_modifiers": {}}, "py/object": "characters.race.Race", "body_template_uid": "humanoid"}, "troll": {"experience_penalty": 0, "uid": "troll", "name": "Troll", "level_tree": {"py/object": "components.level_tree.LevelTree", "abilities_modifiers": {"1": [{"py/object": "components.abilities.ability.Ability", "uid": {"py/object": "components.abilities.power_abilities.PowerAbilities", "py/reduce": [{"py/type": "components.abilities.power_abilities.PowerAbilities"}, {"py/tuple": [1]}, null, null, null]}, "power": 1, "level_progression": 0}]}, "stats_modifiers": {"1": [{"py/object": "components.stats.StatModifier", "uid": {"py/id": 9}, "value": 4, "level_progression": 0}, {"py/object": "components.stats.StatModifier", "uid": {"py/id": 11}, "value": 4, "level_progression": 0}, {"py/object": "components.stats.StatModifier", "uid": {"py/id": 13}, "value": -4, "level_progression": 0}, {"py/object": "components.stats.StatModifier", "uid": {"py/id": 15}, "value": -4, "level_progression": 0}]}}, "py/object": "characters.race.Race", "body_template_uid": "humanoid"}, "orc": {"experience_penalty": 0, "uid": "orc", "name": "Orc", "level_tree": {"py/object": "components.level_tree.LevelTree", "abilities_modifiers": {"1": [{"py/object": "components.abilities.ability.Ability", "uid": {"py/object": "components.abilities.power_abilities.PowerAbilities", "py/reduce": [{"py/type": "components.abilities.power_abilities.PowerAbilities"}, {"py/tuple": [0]}, null, null, null]}, "power": 1, "level_progression": 0}]}, "stats_modifiers": {"1": [{"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["strength"]}, null, null, null]}, "value": 2, "level_progression": 0}, {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["constitution"]}, null, null, null]}, "value": 2, "level_progression": 0}, {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["charisma"]}, null, null, null]}, "value": -2, "level_progression": 0}, {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["intelligence"]}, null, null, null]}, "value": -2, "level_progression": 0}]}}, "py/object": "characters.race.Race", "body_template_uid": "humanoid"}}
+{"troll": {"experience_penalty": 0, "level_tree": {"py/object": "components.level_tree.LevelTree", "stats_modifiers": {"1": [{"py/object": "components.stats.StatModifier", "uid": {"py/id": 9}, "level_progression": 0, "value": 4}, {"py/object": "components.stats.StatModifier", "uid": {"py/id": 11}, "level_progression": 0, "value": 4}, {"py/object": "components.stats.StatModifier", "uid": {"py/id": 13}, "level_progression": 0, "value": -4}, {"py/object": "components.stats.StatModifier", "uid": {"py/id": 15}, "level_progression": 0, "value": -4}]}, "abilities_modifiers": {"1": [{"py/object": "abilities.ability.Ability", "power": 1, "level_progression": 0, "uid": {"py/object": "abilities.power_abilities.PowerAbilities", "py/reduce": [{"py/type": "abilities.power_abilities.PowerAbilities"}, {"py/tuple": [1]}, null, null, null]}}]}}, "py/object": "characters.race.Race", "uid": "troll", "body_template_uid": "humanoid", "name": "Troll"}, "human": {"experience_penalty": 0, "level_tree": {"py/object": "components.level_tree.LevelTree", "stats_modifiers": {}, "abilities_modifiers": {}}, "py/object": "characters.race.Race", "uid": "human", "body_template_uid": "humanoid", "name": "Human"}, "orc": {"experience_penalty": 0, "level_tree": {"py/object": "components.level_tree.LevelTree", "stats_modifiers": {"1": [{"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["strength"]}, null, null, null]}, "level_progression": 0, "value": 2}, {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["constitution"]}, null, null, null]}, "level_progression": 0, "value": 2}, {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["charisma"]}, null, null, null]}, "level_progression": 0, "value": -2}, {"py/object": "components.stats.StatModifier", "uid": {"py/object": "components.stats.Stats", "py/reduce": [{"py/type": "components.stats.Stats"}, {"py/tuple": ["intelligence"]}, null, null, null]}, "level_progression": 0, "value": -2}]}, "abilities_modifiers": {"1": [{"py/object": "abilities.ability.Ability", "power": 1, "level_progression": 0, "uid": {"py/object": "abilities.power_abilities.PowerAbilities", "py/reduce": [{"py/type": "abilities.power_abilities.PowerAbilities"}, {"py/tuple": [0]}, null, null, null]}}]}}, "py/object": "characters.race.Race", "uid": "orc", "body_template_uid": "humanoid", "name": "Orc"}}
