+++ items/item.py
@@ -21,8 +21,7 @@ class Item(object):
     A good loot system can go a long way in terms of extending play time.
     """
     def __init__(self, uid, name="", description="", location=None,
-                 display=None, material=None, stats=None, melee_damage_type=DamageType.Blunt,
-                 wearable_bodyparts_uid=None, worn_layer=0):
+                 display=None, material=None, stats=None, ):
         self.uid = uid
         self._name = name
         self._description = description
@@ -30,9 +29,6 @@ class Item(object):
         self.display = display
         self.material = material
         self.stats = stats
-        self.melee_damage_type = melee_damage_type
-        self.wearable_bodyparts_uid = wearable_bodyparts_uid if wearable_bodyparts_uid else []
-        self.worn_layer = worn_layer
 
         # TODO Items should have weapon stats AND armor stats of D&D
         # TODO They should be affected by material, so you can have a base item of iron breastplate of +2 AC
